+++ api/src/test/java/it/unicam/cs/pa/ma114763/logo/TestUtils.java
@@ -8,7 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
  * @author Lorenzo Lapucci
  */
 public class TestUtils {
-    public static void assertColorStatement(SetColorStatement statement, byte r, byte g, byte b, byte a) {
+    public static void assertColorStatement(SetColorStatement statement, int r, int g, int b, int a) {
         Color c = statement.color();
         assertEquals(r, c.getRed());
         assertEquals(g, c.getGreen());
